% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ByteStream.R
\name{ByteStream}
\alias{ByteStream}
\title{Internal helper class for dealing with byte sequences}
\description{
Internal helper class for dealing with byte sequences

Internal helper class for dealing with byte sequences
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{vec}}{vector of bytes}

\item{\code{idx}}{current index within self$vec
Initialise}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ByteStream-new}{\code{ByteStream$new()}}
\item \href{#method-ByteStream-advance}{\code{ByteStream$advance()}}
\item \href{#method-ByteStream-consume}{\code{ByteStream$consume()}}
\item \href{#method-ByteStream-consume_len2}{\code{ByteStream$consume_len2()}}
\item \href{#method-ByteStream-consume_len4}{\code{ByteStream$consume_len4()}}
\item \href{#method-ByteStream-eos}{\code{ByteStream$eos()}}
\item \href{#method-ByteStream-clone}{\code{ByteStream$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByteStream-new"></a>}}
\if{latex}{\out{\hypertarget{method-ByteStream-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByteStream$new(vec)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vec}}{vector of bytes. will be cast to integer
Advance the indxe pointer without returning the value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByteStream-advance"></a>}}
\if{latex}{\out{\hypertarget{method-ByteStream-advance}{}}}
\subsection{Method \code{advance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByteStream$advance(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{number of bytes to advance
Consume bytes and return the values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByteStream-consume"></a>}}
\if{latex}{\out{\hypertarget{method-ByteStream-consume}{}}}
\subsection{Method \code{consume()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByteStream$consume(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of bytes to consume}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByteStream-consume_len2"></a>}}
\if{latex}{\out{\hypertarget{method-ByteStream-consume_len2}{}}}
\subsection{Method \code{consume_len2()}}{
Consume 2 bytes and interpret as a little-endian integer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByteStream$consume_len2()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByteStream-consume_len4"></a>}}
\if{latex}{\out{\hypertarget{method-ByteStream-consume_len4}{}}}
\subsection{Method \code{consume_len4()}}{
Consume 4 bytes and interpret as a little-endian integer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByteStream$consume_len4()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByteStream-eos"></a>}}
\if{latex}{\out{\hypertarget{method-ByteStream-eos}{}}}
\subsection{Method \code{eos()}}{
Have we reached/exceeded the length of the bytestream
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByteStream$eos()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByteStream-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ByteStream-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByteStream$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
