% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\docType{data}
\name{req}
\alias{req}
\title{Named list functions to generate raw byte vectors to be sent to VICE}
\format{
An object of class \code{list} of length 0.
}
\usage{
req
}
\description{
Functions indexed by both their command code (e.g. "0x31") and 
by their description (e.g. "registers_get")
}
\keyword{datasets}
