% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_advance_instructions}
\alias{req_advance_instructions}
\title{Step over a certain number of instructions.}
\usage{
req_advance_instructions(n, sub1 = TRUE)
}
\arguments{
\item{n}{number of instructions to jump over}

\item{sub1}{Should subroutines count as a single instruction?}
}
\value{
empty response
}
\description{
Step over a certain number of instructions.
}
