% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_autostart}
\alias{req_autostart}
\title{Load a program then return to the monitor}
\usage{
req_autostart(filename, file_idx, run_after_load = TRUE)
}
\arguments{
\item{filename}{Name of file}

\item{file_idx}{If given a disk image, the index of the file to execute.
Default: 0.}

\item{run_after_load}{logical. Default: TRUE}
}
\description{
Load a program then return to the monitor
}
