% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_banks_available}
\alias{req_banks_available}
\title{Gives a listing of all the bank IDs for the running machine with their names.}
\usage{
req_banks_available()
}
\value{
named list of bank IDs
}
\description{
Gives a listing of all the bank IDs for the running machine with their names.
}
