% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_checkpoint_delete}
\alias{req_checkpoint_delete}
\title{Deletes any type of checkpoint. (break, watch, trace)}
\usage{
req_checkpoint_delete(checkpoint)
}
\arguments{
\item{checkpoint}{number 32bit int.}
}
\value{
empty response
}
\description{
Deletes any type of checkpoint. (break, watch, trace)
}
