% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_checkpoint_get}
\alias{req_checkpoint_get}
\title{Gets any type of checkpoint. (break, watch, trace)}
\usage{
req_checkpoint_get(checkpoint)
}
\arguments{
\item{checkpoint}{number 32bit int.}
}
\value{
a named list 
\describe{
  \item{checkpoint}{Checkpoint number}    
  \item{hit}{Currently hit? logical}
  \item{start}{Start address}
  \item{end}{End address}
  \item{stop_when_hit}{Stop when hit? Logical.}
  \item{enabled}{Logical.}
  \item{cpu_op}{CPU operatio: 1 = Load, 2 = Store, 4 = Exec}
  \item{temporary}{Deletes the checkpoint after it has been hit once. 
            This is similar to "until" command, but it will not 
            resume the emulator.}
  \item{hit_count}{Number of hits}
  \item{ignore_count}{Ignore count}
  \item{has_condition}{Has condition? Logical.}
  \item{mem_space}{Memory space.}
}
}
\description{
Gets any type of checkpoint. (break, watch, trace)
}
