% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_checkpoint_list}
\alias{req_checkpoint_list}
\title{List of checkpoints}
\usage{
req_checkpoint_list()
}
\value{
causes VICE to emit a series of responses (as would be returned
by a sequence of \code{req_checkpoint_get()}) followed by the actual
response body which is just a count of the total number of checkpoints
}
\description{
List of checkpoints
}
