% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_checkpoint_set}
\alias{req_checkpoint_set}
\title{Sets any type of checkpoint.}
\usage{
req_checkpoint_set(
  start,
  end,
  stop_when_hit = TRUE,
  enabled = TRUE,
  cpu_op = 1,
  temporary = TRUE,
  mem_space = machine$mem_space$main
)
}
\arguments{
\item{start}{start address}

\item{end}{end address}

\item{stop_when_hit}{logical. Default: TRUE}

\item{enabled}{logical. default: TRUE}

\item{cpu_op}{1 = load. 2 = store. 4 = exec. devault: 1}

\item{temporary}{Deletes the checkpoint after it has been hit once. 
This is similar to "until" command, but it will not resume the emulator.}

\item{mem_space}{memory space (integer). Default: 0 (main memory). Other
acceptable values 1=drive8, 2=drive9, 3=drive10, 4=drive11}
}
\value{
checkpoint list. Samve as \code{req_checkpont_set}
}
\description{
This combines the functionality of several 
textual commands (break, watch, # trace) into one, as they are all the 
same with only minor variations. To set conditions, see section 13.4.8 
Condition set (0x22) after executing this one.
}
