% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_checkpoint_toggle}
\alias{req_checkpoint_toggle}
\title{Checkpoint toggle}
\usage{
req_checkpoint_toggle(checkpoint, enabled)
}
\arguments{
\item{checkpoint}{number 32bit int.}

\item{enabled}{Should checkpoint be enabled? logical.}
}
\value{
empty response
}
\description{
Checkpoint toggle
}
