% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_condition_set}
\alias{req_condition_set}
\title{Sets a condition on an existing checkpoint. It is not currently possible to 
retrieve conditions after setting them.}
\usage{
req_condition_set(checkpoint, condition_expr)
}
\arguments{
\item{checkpoint}{number 32bit int.}

\item{condition_expr}{Condition expression string. This is the same format 
used on the command line.}
}
\value{
empty response
}
\description{
Sets a condition on an existing checkpoint. It is not currently possible to 
retrieve conditions after setting them.
}
