% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_dump}
\alias{req_dump}
\title{Saves the machine state to a file.}
\usage{
req_dump(filename, save_roms = FALSE, save_disks = FALSE)
}
\arguments{
\item{filename}{filename to save machine state into}

\item{save_roms}{save ROMs to snapshot file?  Logical. default: FALSE}

\item{save_disks}{save disks to snapshot file. Logical. default: FALSE}
}
\value{
empty response
}
\description{
Saves the machine state to a file.
}
