% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_execute_until_return}
\alias{req_execute_until_return}
\title{Continues execution and returns to the monitor just after the next RTS or RTI is executed.}
\usage{
req_execute_until_return()
}
\value{
empty response
}
\description{
This command is the same as "return" in the text monitor.
}
