% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_joyport_set}
\alias{req_joyport_set}
\title{Set the simulated joyport value.}
\usage{
req_joyport_set(port, value)
}
\arguments{
\item{port}{the port to set the value on}

\item{value}{value to set}
}
\value{
empty response
}
\description{
Set the simulated joyport value.
}
