% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_keyboard_feed}
\alias{req_keyboard_feed}
\title{Add text to the keyboard buffer.}
\usage{
req_keyboard_feed(text)
}
\arguments{
\item{text}{PETSCII text}
}
\value{
empty response
}
\description{
Add text to the keyboard buffer.
}
