% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_memory_get}
\alias{req_memory_get}
\title{Reads a chunk of memory from a start address to an end address (inclusive).}
\usage{
req_memory_get(start, end, mem_space = 0, side_effects = FALSE, bank_id = 0)
}
\arguments{
\item{start}{start address (integer)}

\item{end}{end address (integer)}

\item{mem_space}{memory space (integer). Default: 0 (main memory). Other
acceptable values 1=drive8, 2=drive9, 3=drive10, 4=drive11}

\item{side_effects}{Should the read cause side effects? Default: FALSE}

\item{bank_id}{which bank you want. This is dependent on your machine. 
If the memspace 
selected doesn't support banks, this value is ignored.}
}
\value{
vector of raw bytes
}
\description{
13.4.1 Memory get (0x01)
}
\details{
cmd  <- 0x01
}
\seealso{
[req_banks_available()]
}
