% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_memory_set}
\alias{req_memory_set}
\title{Writes a chunk of memory from a start address to an end address (inclusive).}
\usage{
req_memory_set(
  bytes,
  start,
  bank_id = 0,
  mem_space = machine$mem_space$main,
  side_effects = FALSE
)
}
\arguments{
\item{bytes}{raw vector of bytes to be written to memory}

\item{start}{start address (integer)}

\item{bank_id}{which bank you want. This is dependent on your machine. 
If the memspace 
selected doesn't support banks, this value is ignored.}

\item{mem_space}{memory space (integer). Default: 0 (main memory). Other
acceptable values 1=drive8, 2=drive9, 3=drive10, 4=drive11}

\item{side_effects}{Should the read cause side effects? Default: FALSE}
}
\value{
empty response
}
\description{
Writes a chunk of memory from a start address to an end address (inclusive).
}
