% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_palette_get}
\alias{req_palette_get}
\title{Get the colors in the current palette}
\usage{
req_palette_get()
}
\value{
character vector of colours
}
\description{
Get the colors in the current palette
}
