% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_registers_available}
\alias{req_registers_available}
\title{Gives a listing of all the registers for the running machine with their names.}
\usage{
req_registers_available(mem_space = machine$mem_space$main)
}
\arguments{
\item{mem_space}{memory space (integer). Default: 0 (main memory). Other
acceptable values 1=drive8, 2=drive9, 3=drive10, 4=drive11}
}
\value{
named list of regiseters
}
\description{
Gives a listing of all the registers for the running machine with their names.
}
