% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_registers_get}
\alias{req_registers_get}
\title{Get details about the registers}
\usage{
req_registers_get(mem_space = machine$mem_space$main)
}
\arguments{
\item{mem_space}{memory space (integer). Default: 0 (main memory). Other
acceptable values 1=drive8, 2=drive9, 3=drive10, 4=drive11}
}
\value{
named integer vector where names are the register names.
}
\description{
Get details about the registers
}
