% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_registers_set}
\alias{req_registers_set}
\title{Set the register values}
\usage{
req_registers_set(..., mem_space = machine$mem_space$main)
}
\arguments{
\item{...}{name/value pairs specifying values for the named registers.
e.g. \code{req_register_set(A = 12)}}

\item{mem_space}{memory space (integer). Default: 0 (main memory). Other
acceptable values 1=drive8, 2=drive9, 3=drive10, 4=drive11}
}
\value{
named integer vector where names are the register names.
}
\description{
Set the register values
}
