% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_reset}
\alias{req_reset}
\title{Reset the system or a drive}
\usage{
req_reset(hard_reset = TRUE, drive = NULL)
}
\arguments{
\item{hard_reset}{logical. default: TRUE}

\item{drive}{integer. 8-11. If set, then this drive is reset rather than 
the main system.}
}
\description{
Reset the system or a drive
}
