% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_resource_set}
\alias{req_resource_set}
\title{Set a resource value in the emulator. See section 6.1 Format of resource files.}
\usage{
req_resource_set(resource, value)
}
\arguments{
\item{resource}{resource name. See \url{https://vice-emu.sourceforge.io/vice_6.html#SEC84}}

\item{value}{value to set. allowed: string or integer.}
}
\value{
empty response
}
\description{
Set a resource value in the emulator. See section 6.1 Format of resource files.
}
