% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_undump}
\alias{req_undump}
\title{Loads the machine state from a file.}
\usage{
req_undump(filename)
}
\arguments{
\item{filename}{filename containing dump of the machine state.}
}
\value{
The current program counter position
}
\description{
Loads the machine state from a file.
}
