% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{req_vice_info}
\alias{req_vice_info}
\title{Get general information about VICE. Currently returns the versions.}
\usage{
req_vice_info()
}
\value{
list with vice and SVN versions
}
\description{
Get general information about VICE. Currently returns the versions.
}
