% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{run_prg}
\alias{run_prg}
\title{Run a PRG given as a numeric vector or a filename}
\usage{
run_prg(prg, ...)
}
\arguments{
\item{prg}{filename or raw/numeric vector}

\item{...}{extra arguments passed to \code{send_req()}}
}
\description{
Run a PRG given as a numeric vector or a filename
}
