% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_screenshot}
\alias{save_screenshot}
\title{Take a screenshot and save as PNG file}
\usage{
save_screenshot(filename, scale = 4, keep_border = TRUE)
}
\arguments{
\item{filename}{PNG filename}

\item{scale}{scale factor applied to image before saving. defualt: 4}

\item{keep_border}{keep the borders as part of the screenshot.}
}
\description{
Take a screenshot and save as PNG file
}
