% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{send_req}
\alias{send_req}
\title{Deliver a request of raw bytes to the VICE binary monitor}
\usage{
send_req(
  request,
  parse_response = TRUE,
  keep_raw = FALSE,
  wait = 0.2,
  host = "localhost",
  port = 6502
)
}
\arguments{
\item{request}{raw vector}

\item{parse_response}{default: TRUE. If FALSE then return raw vector}

\item{keep_raw}{should the raw bytes be returned as part of the parsed response?
Default: FALSE}

\item{wait}{how long to wait before checking for the response to the request. 
Default: 0.2s.  Different commands, networks and Operating Systems will have
different requirements here.}

\item{host}{host to connect to. Default: "localhost"}

\item{port}{port on host to connect to. Default: 6502.  This is the default
port set by VICE when running the binary monitor.}
}
\value{
named list with parsed response or a raw vector (if \code{parse_response = FALSE})
}
\description{
Deliver a request of raw bytes to the VICE binary monitor
}
