% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{take_screenshot}
\alias{take_screenshot}
\title{Take a screenshot}
\usage{
take_screenshot(keep_border = TRUE)
}
\arguments{
\item{keep_border}{keep the borders as part of the screenshot.}
}
\value{
raster
}
\description{
Take a screenshot
}
